pega-adk/requirements.txt
# ADK Agent Dependencies
google-adk>=1.2.0
